[Patch v4] Support transforming inherited methods
This makes it possible to transform objects which are instances of a
class without knowing which methods are present on the instance object
and which are inherited through the prototype chain.  Include/Exclude is
handled the same as for non-inherited methods.

In addition to transforming inherited methods, when the inherited option
is enabled the result object is created with the prototype of the
original object so that it behaves the same as the original object for
instanceof checks.

Fixes #14

Changes in v2:
- Default-initialize ret to an empty object.

Changes in v3:
- Replace reduction over keys array with a for-in loop.

Changes in v4:
- Rebase on to latest master (ES2015, ava changes, style changes).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>